# Added wildcard feature to .bazelignore