--- a/src/main/java/com/google/devtools/build/lib/skyframe/IgnoredPackagePrefixesValue.java
+++ b/src/main/java/com/google/devtools/build/lib/skyframe/IgnoredPackagePrefixesValue.java
@@ -16,31 +16,66 @@ package com.google.devtools.build.lib.skyframe;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Interner;
+import com.google.common.collect.Maps;
+import com.google.common.collect.Sets;
 import com.google.devtools.build.lib.cmdline.RepositoryName;
 import com.google.devtools.build.lib.concurrent.BlazeInterners;
 import com.google.devtools.build.lib.skyframe.serialization.autocodec.AutoCodec;
+import com.google.devtools.build.lib.vfs.Dirent;
+import com.google.devtools.build.lib.vfs.Path;
 import com.google.devtools.build.lib.vfs.PathFragment;
+import com.google.devtools.build.lib.vfs.Root;
+import com.google.devtools.build.lib.vfs.RootedPath;
+import com.google.devtools.build.lib.vfs.Symlinks;
+import com.google.devtools.build.lib.vfs.UnixGlob;
 import com.google.devtools.build.skyframe.AbstractSkyKey;
 import com.google.devtools.build.skyframe.SkyFunctionName;
 import com.google.devtools.build.skyframe.SkyKey;
 import com.google.devtools.build.skyframe.SkyValue;
+import java.util.Collection;
+import java.util.Map;
+import java.util.Set;
+import java.util.Stack;
+import java.util.regex.Pattern;
 
 /** An immutable set of package name prefixes that should be ignored. */
 @AutoCodec
 public class IgnoredPackagePrefixesValue implements SkyValue {
-  private final ImmutableSet<PathFragment> patterns;
+  private final Set<PathFragment> ignoredPackagePrefixes;
+  private final ImmutableSet<PathFragment> bazelignoreEntries;
+  private final ImmutableSet<String> wildcardPatterns;
+  private final Set<PathFragment> visitedPackage;
+  private final Root workspaceRoot;
+  private final Map<String, Pattern> patternCache;
 
   @AutoCodec @AutoCodec.VisibleForSerialization
   public static final IgnoredPackagePrefixesValue EMPTY_LIST =
-      new IgnoredPackagePrefixesValue(ImmutableSet.of());
+          new IgnoredPackagePrefixesValue(ImmutableSet.of(), null);
 
-  private IgnoredPackagePrefixesValue(ImmutableSet<PathFragment> patterns) {
-    this.patterns = Preconditions.checkNotNull(patterns);
+  private IgnoredPackagePrefixesValue(ImmutableSet<PathFragment> patterns, Root workspaceRoot) {
+    bazelignoreEntries = Preconditions.checkNotNull(patterns);
+
+    ImmutableSet.Builder<String> wildcardPatternsBuilder = ImmutableSet.builder();
+    ignoredPackagePrefixes = Sets.newConcurrentHashSet();
+
+    for (PathFragment pattern: patterns) {
+      String path = pattern.getPathString();
+      if (UnixGlob.isWildcardFree(path)) {
+        ignoredPackagePrefixes.add(pattern);
+      } else {
+        wildcardPatternsBuilder.add(path);
+      }
+    }
+
+    wildcardPatterns = wildcardPatternsBuilder.build();
+    patternCache = Maps.newHashMap();
+    visitedPackage = Sets.newHashSet();
+    this.workspaceRoot = workspaceRoot;
   }
 
   @AutoCodec.Instantiator
-  public static IgnoredPackagePrefixesValue of(ImmutableSet<PathFragment> patterns) {
-    return patterns.isEmpty() ? EMPTY_LIST : new IgnoredPackagePrefixesValue(patterns);
+  public static IgnoredPackagePrefixesValue of(ImmutableSet<PathFragment> patterns, Root workspaceRoot) {
+    return patterns.isEmpty() ? EMPTY_LIST : new IgnoredPackagePrefixesValue(patterns, workspaceRoot);
   }
 
   /** Creates a key from the main repository. */
@@ -54,19 +89,35 @@ public class IgnoredPackagePrefixesValue implements SkyValue {
   }
 
   public ImmutableSet<PathFragment> getPatterns() {
-    return patterns;
+    return ImmutableSet.copyOf(ignoredPackagePrefixes);
+  }
+
+  public ImmutableSet<PathFragment> getPatterns(PathFragment base) {
+    if (workspaceRoot == null || visitedPackage.contains(base)) {
+      return ImmutableSet.copyOf(ignoredPackagePrefixes);
+    }
+    Path root = RootedPath.toRootedPath(workspaceRoot, base).asPath();
+    visitedPackage.add(base);
+    for (String pattern: wildcardPatterns) {
+      if (UnixGlob.matches(pattern, base.getPathString(), patternCache)) {
+        ignoredPackagePrefixes.add(base);
+        return ImmutableSet.copyOf(ignoredPackagePrefixes);
+      }
+    }
+    collectIgnoredDirectories(root);
+    return ImmutableSet.copyOf(ignoredPackagePrefixes);
   }
 
   @Override
   public int hashCode() {
-    return patterns.hashCode();
+    return bazelignoreEntries.hashCode();
   }
 
   @Override
   public boolean equals(Object obj) {
     if (obj instanceof IgnoredPackagePrefixesValue) {
       IgnoredPackagePrefixesValue other = (IgnoredPackagePrefixesValue) obj;
-      return this.patterns.equals(other.patterns);
+      return this.bazelignoreEntries.equals(other.bazelignoreEntries);
     }
     return false;
   }
@@ -89,6 +140,47 @@ public class IgnoredPackagePrefixesValue implements SkyValue {
     @Override
     public SkyFunctionName functionName() {
       return SkyFunctions.IGNORED_PACKAGE_PREFIXES;
+    }
+  }
+
+  /**
+   * Scan directories iteratively under {@code rootDirectory} and match them to avaliable patterns
+   * in {@code wildcardPatterns}. Add the matched directories to {@code ignoredPackagePrefixesBuilder}.
+   *
+   * @param rootDirectory the directory of package
+   */
+  private void collectIgnoredDirectories(Path rootDirectory) {
+    Stack<Path> rootStack = new Stack<>();
+    rootStack.push(rootDirectory);
+    Stack<String> relativeStack = new Stack<>();
+    relativeStack.push("");
+
+    while (!rootStack.isEmpty()) {
+      Path root = rootStack.pop();
+      String relative = relativeStack.pop();
+      Collection<Dirent> dirents;
+      try {
+        dirents = root.readdir(Symlinks.FOLLOW);
+      } catch (Exception ignored) {
+        continue;
+      }
+
+      nextMatch:
+      for (Dirent dirent : dirents) {
+        String childRelative = relative.isEmpty()
+                ? dirent.getName()
+                : relative + "/" + dirent.getName();
+        if (dirent.getType() == Dirent.Type.DIRECTORY) {
+          for (String pattern: wildcardPatterns) {
+            if (UnixGlob.matches(pattern, childRelative, patternCache)) {
+              ignoredPackagePrefixes.add(PathFragment.create(childRelative));
+              continue nextMatch;
+            }
+          }
+          rootStack.push(root.getChild(dirent.getName()));
+          relativeStack.push(childRelative);
+        }
+      }
     }
   }
 }