This pr add wildcard features to .bazelignore. It aims to fix Issue:https://github.com/bazelbuild/bazel/issues/7093. The supported wildcard syntax is same as those in `UnixGlob` (`?`, `*`, and `**`).

**Changes**
I modified `IgnoredPackagePrefixesValue` so that it will evaluate ignored path prefixes every time new package is look up. It done that by globing all the directories under the package (including the package) and add the directories with path matches an ignored pattern to the set of ignored prefixes. The globing will not continue if the current directory is ignored (early stop). The globing logic is similar to `scanDirectoryRecursively` in `BinTool` expect it use an iterative approach instead of recursion. 

The changes are implemented in the `IgnoredPackagePrefixesValue` rather than `IgnoredPackagePrefixesFunction` because `IgnoredPackagePrefixesFunction` will not re-evaluate ignored prefixes when new directories are added, which cause them not being ignored even they match a wildcard pattern in `.bazelignore`.

I also added 2 tests to bazelignore_test.sh. The `test_new_directory_is_ignored` check will newly added directories be ignored and `test_does_not_glob_into_ignored_directory_with_wildcard` check can wildcard entry successfully ignore directories.

**Impact**
The following are metrics by running the command in a large android repository with around 560 Gradle modules and we want to exclude all build folder for our Bazel build. Here is the performance metrics before and after this change.
```
bazel shutdown
time bazel query //src:main
```
**After changes**
Without `**/build` (no modification in `.bazelignore`):
```
bazel query //src:main 0.82s user 4.58s system 33% cpu 16.289 total
bazel query //src:main  0.68s user 1.29s system 21% cpu 9.149 total
bazel query //src:main  0.70s user 1.38s system 21% cpu 9.819 total
```

With `**/build` (replace all prefixes ending with `/build` (around 920 prefixes) to `**/build` in `.bazelignore`):
```
bazel query //src:main  0.74s user 1.28s system 22% cpu 9.121 total
bazel query //src:main  0.74s user 1.30s system 24% cpu 8.220 total
bazel query //src:main  0.73s user 1.26s system 23% cpu 8.457 total
```
**Original**
```
bazel query //src:main  0.66s user 1.18s system 28% cpu 6.390 total
bazel query //src:main  0.68s user 1.20s system 27% cpu 6.844 total
bazel query //src:main  0.66s user 1.18s system 28% cpu 6.583 total
```