@nkoroste 2021-09-25
@janakdr any updates on this? we would love to use this feature as it significantly improves the IDE experience in a repo that has both Gradle and Bazel since it allows you to ignore all auto generated Gradle build/ dirs