@nkoroste 2021-09-27
Thanks for the update, I wasn't aware of the other PR https://github.com/bazelbuild/bazel/pull/13807 I will subscribe to it instead