@sgowroji 2022-04-22
Hello @David-Lam, I was going through the recent comments from [here](https://github.com/bazelbuild/bazel/pull/13756#issuecomment-927984753). Can you please share the latest update on this PR. Thanks!