@David-Lam 2022-04-22
> Hello @David-Lam, I was going through the recent comments from [here](https://github.com/bazelbuild/bazel/pull/13756#issuecomment-927984753). Can you please share the latest update on this PR. Thanks!

Yes, I am currently trying to change all reference to the set of ignored prefixes the to a custom class. My progress is very slow since the ignored prefixes is passed to many places in the project. Feel free to work on the issue if you have new idea!