@nkoroste 2022-11-08
fyi - I think the newer version of this PR is here https://github.com/bazelbuild/bazel/pull/13807